Start moving the LED Driver into a library
