+++ leddriver/include/leddriver/pioCommunicator.hpp
@@ -1,0 +1,19 @@
+#pragma once
+
+#include "hardware/pio.h"
+
+namespace LEDDriver
+{
+    class PIOCommunicator
+    {
+    public:
+        PIO pio;
+        uint sm;
+        uint pin_clk;
+        uint pin_data;
+
+        void write32blocking(const uint32_t *src, const size_t n);
+
+        void waitTXdrain();
+    };
+}
